更新代码

+++ lyn-customer/customer-api/src/main/java/com/lyn/customer/api/service/CustomerService.java
@@ -6,5 +6,5 @@ public interface CustomerService {
 
     String test(String name);
 
-    CustomerInfo findCustomerInfoByPrimary(long customerId) throws Exception;
+    CustomerInfo findCustomerInfoByPrimary(Integer customerId) throws Exception;
 }
